@@ -42,34 +46,38 @@
   },
   "bundleDependencies": false,
   "dependencies": {
-    "find-cache-dir": "^0.1.1",
-    "loader-utils": "^0.2.16",
-    "mkdirp": "^0.5.1",
-    "object-assign": "^4.0.1"
+    "find-cache-dir": "^1.0.0",
+    "loader-utils": "^1.0.2",
+    "mkdirp": "^0.5.1"
   },
   "deprecated": false,
   "description": "babel module loader for webpack",
   "devDependencies": {
-    "ava": "^0.17.0",
+    "ava": "^0.23.0",
     "babel-cli": "^6.18.0",
     "babel-core": "^6.0.0",
-    "babel-eslint": "^7.1.0",
-    "babel-plugin-istanbul": "^3.0.0",
+    "babel-eslint": "^8.0.0",
+    "babel-plugin-istanbul": "^4.0.0",
     "babel-plugin-react-intl": "^2.1.3",
-    "babel-preset-es2015": "^6.0.0",
-    "babel-preset-latest": "^6.16.0",
+    "babel-preset-env": "^1.2.0",
     "babel-register": "^6.18.0",
-    "codecov": "^1.0.1",
-    "cross-env": "^2.0.0",
-    "eslint": "^3.8.1",
-    "eslint-config-babel": "^6.0.0",
+    "cross-env": "^5.0.0",
+    "eslint": "^4.1.0",
+    "eslint-config-babel": "^7.0.0",
     "eslint-plugin-flowtype": "^2.25.0",
-    "nyc": "^10.0.0",
-    "react": "^15.1.0",
+    "eslint-plugin-prettier": "^2.1.2",
+    "husky": "^0.14.0",
+    "lint-staged": "^4.0.0",
+    "nyc": "^11.0.1",
+    "prettier": "^1.2.2",
+    "react": "^16.0.0",
     "react-intl": "^2.1.2",
     "react-intl-webpack-plugin": "^0.0.3",
     "rimraf": "^2.4.3",
-    "webpack": "^2.2.0-rc"
+    "webpack": "^4.0.0"
+  },
+  "engines": {
+    "node": ">=4"
   },
   "files": [
     "lib"
@@ -98,8 +132,8 @@
     "instrument": false
   },
   "peerDependencies": {
-    "babel-core": "^6.0.0",
-    "webpack": "1 || 2 || ^2.1.0-beta || ^2.2.0-rc"
+    "babel-core": "6",
+    "webpack": "2 || 3 || 4"
   },
   "repository": {
     "type": "git",
