@@ -1,27 +1,34 @@
 {
-  "_from": "globby@^6.1.0",
-  "_id": "globby@6.1.0",
+  "_args": [
+    [
+      "globby@8.0.1",
+      "/Users/mattharris/dscovr"
+    ]
+  ],
+  "_development": true,
+  "_from": "globby@8.0.1",
+  "_id": "globby@8.0.1",
   "_inBundle": false,
-  "_integrity": "sha1-9abXDoOV4hyFj7BInWTfAkJNUGw=",
+  "_integrity": "sha512-oMrYrJERnKBLXNLVTqhm3vPEdJ/b2ZE28xN4YARiix1NOIOBPEpOUnm844K1iu/BkphCaf2WNFwMszv8Soi1pw==",
   "_location": "/globby",
   "_phantomChildren": {},
   "_requested": {
-    "type": "range",
+    "type": "version",
     "registry": true,
-    "raw": "globby@^6.1.0",
+    "raw": "globby@8.0.1",
     "name": "globby",
     "escapedName": "globby",
-    "rawSpec": "^6.1.0",
+    "rawSpec": "8.0.1",
     "saveSpec": null,
-    "fetchSpec": "^6.1.0"
+    "fetchSpec": "8.0.1"
   },
   "_requiredBy": [
-    "/del"
+    "/mem-fs-editor",
+    "/yeoman-environment"
   ],
-  "_resolved": "https://registry.npmjs.org/globby/-/globby-6.1.0.tgz",
-  "_shasum": "f5a6d70e8395e21c858fb0489d64df02424d506c",
-  "_spec": "globby@^6.1.0",
-  "_where": "/Users/mattharris/dscovr/node_modules/del",
+  "_resolved": "https://registry.npmjs.org/globby/-/globby-8.0.1.tgz",
+  "_spec": "8.0.1",
+  "_where": "/Users/mattharris/dscovr",
   "author": {
     "name": "Sindre Sorhus",
     "email": "sindresorhus@gmail.com",
@@ -30,29 +37,30 @@
   "bugs": {
     "url": "https://github.com/sindresorhus/globby/issues"
   },
-  "bundleDependencies": false,
   "dependencies": {
     "array-union": "^1.0.1",
-    "glob": "^7.0.3",
-    "object-assign": "^4.0.1",
-    "pify": "^2.0.0",
-    "pinkie-promise": "^2.0.0"
+    "dir-glob": "^2.0.0",
+    "fast-glob": "^2.0.2",
+    "glob": "^7.1.2",
+    "ignore": "^3.3.5",
+    "pify": "^3.0.0",
+    "slash": "^1.0.0"
   },
-  "deprecated": false,
   "description": "Extends `glob` with support for multiple patterns and exposes a Promise API",
   "devDependencies": {
     "ava": "*",
-    "glob-stream": "github:gulpjs/glob-stream#master",
+    "glob-stream": "^6.1.0",
     "globby": "github:sindresorhus/globby#master",
     "matcha": "^0.7.0",
     "rimraf": "^2.2.8",
-    "xo": "^0.16.0"
+    "xo": "^0.18.0"
   },
   "engines": {
-    "node": ">=0.10.0"
+    "node": ">=4"
   },
   "files": [
-    "index.js"
+    "index.js",
+    "gitignore.js"
   ],
   "homepage": "https://github.com/sindresorhus/globby#readme",
   "keywords": [
@@ -85,7 +93,9 @@
     "utility",
     "wildcard",
     "wildcards",
-    "promise"
+    "promise",
+    "gitignore",
+    "git"
   ],
   "license": "MIT",
   "name": "globby",
@@ -94,8 +104,8 @@
     "url": "git+https://github.com/sindresorhus/globby.git"
   },
   "scripts": {
-    "bench": "npm update glob-stream && matcha bench.js",
+    "bench": "npm update glob-stream fast-glob && matcha bench.js",
     "test": "xo && ava"
   },
-  "version": "6.1.0"
+  "version": "8.0.1"
 }
