@@ -1,29 +1,35 @@
 {
   "_args": [
     [
-      "statuses@https://registry.npmjs.org/statuses/-/statuses-1.3.1.tgz",
+      "statuses@1.4.0",
       "/Users/mattharris/dscovr"
     ]
   ],
-  "_from": "statuses@https://registry.npmjs.org/statuses/-/statuses-1.3.1.tgz",
-  "_id": "statuses@https://registry.npmjs.org/statuses/-/statuses-1.3.1.tgz",
+  "_from": "statuses@1.4.0",
+  "_id": "statuses@1.4.0",
   "_inBundle": false,
-  "_integrity": "sha1-+vUbnrdKrvOzrPStX2Gr8ky3uT4=",
+  "_integrity": "sha512-zhSCtt8v2NDrRlPQpCNtw/heZLtfUDqxBM1udqikb/Hbk52LK4nQSwr10u77iopCW5LsyHpuXS0GnEc48mLeew==",
   "_location": "/statuses",
   "_phantomChildren": {},
   "_requested": {
-    "type": "remote",
-    "raw": "statuses@https://registry.npmjs.org/statuses/-/statuses-1.3.1.tgz",
+    "type": "version",
+    "registry": true,
+    "raw": "statuses@1.4.0",
     "name": "statuses",
     "escapedName": "statuses",
-    "rawSpec": "https://registry.npmjs.org/statuses/-/statuses-1.3.1.tgz",
-    "saveSpec": "https://registry.npmjs.org/statuses/-/statuses-1.3.1.tgz",
-    "fetchSpec": "https://registry.npmjs.org/statuses/-/statuses-1.3.1.tgz"
+    "rawSpec": "1.4.0",
+    "saveSpec": null,
+    "fetchSpec": "1.4.0"
   },
   "_requiredBy": [
-    "/raw-body/http-errors"
+    "/express",
+    "/finalhandler",
+    "/http-errors",
+    "/raw-body/http-errors",
+    "/send"
   ],
-  "_spec": "https://registry.npmjs.org/statuses/-/statuses-1.3.1.tgz",
+  "_resolved": "https://registry.npmjs.org/statuses/-/statuses-1.4.0.tgz",
+  "_spec": "1.4.0",
   "_where": "/Users/mattharris/dscovr",
   "bugs": {
     "url": "https://github.com/jshttp/statuses/issues"
@@ -41,13 +47,17 @@
   ],
   "description": "HTTP status utility",
   "devDependencies": {
-    "csv-parse": "1.1.7",
-    "eslint": "3.10.0",
-    "eslint-config-standard": "6.2.1",
-    "eslint-plugin-promise": "3.3.2",
-    "eslint-plugin-standard": "2.0.1",
+    "csv-parse": "1.2.4",
+    "eslint": "3.19.0",
+    "eslint-config-standard": "10.2.1",
+    "eslint-plugin-import": "2.8.0",
+    "eslint-plugin-markdown": "1.0.0-beta.6",
+    "eslint-plugin-node": "5.2.0",
+    "eslint-plugin-promise": "3.6.0",
+    "eslint-plugin-standard": "3.0.1",
     "istanbul": "0.4.5",
     "mocha": "1.21.5",
+    "raw-body": "2.3.2",
     "stream-to-array": "2.3.0"
   },
   "engines": {
@@ -73,12 +83,12 @@
   },
   "scripts": {
     "build": "node scripts/build.js",
-    "fetch": "node scripts/fetch.js",
-    "lint": "eslint .",
+    "fetch": "node scripts/fetch-apache.js && node scripts/fetch-iana.js && node scripts/fetch-nginx.js && node scripts/fetch-node.js",
+    "lint": "eslint --plugin markdown --ext js,md .",
     "test": "mocha --reporter spec --check-leaks --bail test/",
     "test-ci": "istanbul cover node_modules/mocha/bin/_mocha --report lcovonly -- --reporter spec --check-leaks test/",
     "test-cov": "istanbul cover node_modules/mocha/bin/_mocha -- --reporter dot --check-leaks test/",
     "update": "npm run fetch && npm run build"
   },
-  "version": "https://registry.npmjs.org/statuses/-/statuses-1.3.1.tgz"
+  "version": "1.4.0"
 }
