@@ -1,27 +1,33 @@
 {
-  "_from": "find-up@^1.0.0",
-  "_id": "find-up@1.1.2",
+  "_args": [
+    [
+      "find-up@2.1.0",
+      "/Users/mattharris/dscovr"
+    ]
+  ],
+  "_development": true,
+  "_from": "find-up@2.1.0",
+  "_id": "find-up@2.1.0",
   "_inBundle": false,
-  "_integrity": "sha1-ay6YIrGizgpgq2TWEOzK1TyyTQ8=",
+  "_integrity": "sha1-RdG35QbHF93UgndaK3eSCjwMV6c=",
   "_location": "/webpack-dev-server/find-up",
   "_phantomChildren": {},
   "_requested": {
-    "type": "range",
+    "type": "version",
     "registry": true,
-    "raw": "find-up@^1.0.0",
+    "raw": "find-up@2.1.0",
     "name": "find-up",
     "escapedName": "find-up",
-    "rawSpec": "^1.0.0",
+    "rawSpec": "2.1.0",
     "saveSpec": null,
-    "fetchSpec": "^1.0.0"
+    "fetchSpec": "2.1.0"
   },
   "_requiredBy": [
-    "/webpack-dev-server/read-pkg-up"
+    "/webpack-dev-server/yargs"
   ],
-  "_resolved": "https://registry.npmjs.org/find-up/-/find-up-1.1.2.tgz",
-  "_shasum": "6b2e9822b1a2ce0a60ab64d610eccad53cb24d0f",
-  "_spec": "find-up@^1.0.0",
-  "_where": "/Users/mattharris/dscovr/node_modules/webpack-dev-server/node_modules/read-pkg-up",
+  "_resolved": "https://registry.npmjs.org/find-up/-/find-up-2.1.0.tgz",
+  "_spec": "2.1.0",
+  "_where": "/Users/mattharris/dscovr",
   "author": {
     "name": "Sindre Sorhus",
     "email": "sindresorhus@gmail.com",
@@ -30,12 +36,9 @@
   "bugs": {
     "url": "https://github.com/sindresorhus/find-up/issues"
   },
-  "bundleDependencies": false,
   "dependencies": {
-    "path-exists": "^2.0.0",
-    "pinkie-promise": "^2.0.0"
+    "locate-path": "^2.0.0"
   },
-  "deprecated": false,
   "description": "Find a file by walking up parent directories",
   "devDependencies": {
     "ava": "*",
@@ -43,7 +46,7 @@
     "xo": "*"
   },
   "engines": {
-    "node": ">=0.10.0"
+    "node": ">=4"
   },
   "files": [
     "index.js"
@@ -79,5 +82,8 @@
   "scripts": {
     "test": "xo && ava"
   },
-  "version": "1.1.2"
+  "version": "2.1.0",
+  "xo": {
+    "esnext": true
+  }
 }
