@@ -1,50 +1,53 @@
 {
-  "_from": "yargs@6.6.0",
-  "_id": "yargs@6.6.0",
+  "_args": [
+    [
+      "yargs@11.0.0",
+      "/Users/mattharris/dscovr"
+    ]
+  ],
+  "_development": true,
+  "_from": "yargs@11.0.0",
+  "_id": "yargs@11.0.0",
   "_inBundle": false,
-  "_integrity": "sha1-eC7CHvQDNF+DCoCMo9UTr1YGUgg=",
+  "_integrity": "sha512-Rjp+lMYQOWtgqojx1dEWorjCofi1YN7AoFvYV7b1gx/7dAAeuI4kN5SZiEvr0ZmsZTOpDRcCqrpI10L31tFkBw==",
   "_location": "/webpack-dev-server/yargs",
   "_phantomChildren": {},
   "_requested": {
     "type": "version",
     "registry": true,
-    "raw": "yargs@6.6.0",
+    "raw": "yargs@11.0.0",
     "name": "yargs",
     "escapedName": "yargs",
-    "rawSpec": "6.6.0",
+    "rawSpec": "11.0.0",
     "saveSpec": null,
-    "fetchSpec": "6.6.0"
+    "fetchSpec": "11.0.0"
   },
   "_requiredBy": [
     "/webpack-dev-server"
   ],
-  "_resolved": "https://registry.npmjs.org/yargs/-/yargs-6.6.0.tgz",
-  "_shasum": "782ec21ef403345f830a808ca3d513af56065208",
-  "_spec": "yargs@6.6.0",
-  "_where": "/Users/mattharris/dscovr/node_modules/webpack-dev-server",
+  "_resolved": "https://registry.npmjs.org/yargs/-/yargs-11.0.0.tgz",
+  "_spec": "11.0.0",
+  "_where": "/Users/mattharris/dscovr",
   "bugs": {
     "url": "https://github.com/yargs/yargs/issues"
   },
-  "bundleDependencies": false,
   "dependencies": {
-    "camelcase": "^3.0.0",
-    "cliui": "^3.2.0",
+    "cliui": "^4.0.0",
     "decamelize": "^1.1.1",
+    "find-up": "^2.1.0",
     "get-caller-file": "^1.0.1",
-    "os-locale": "^1.4.0",
-    "read-pkg-up": "^1.0.1",
+    "os-locale": "^2.0.0",
     "require-directory": "^2.1.1",
     "require-main-filename": "^1.0.1",
     "set-blocking": "^2.0.0",
-    "string-width": "^1.0.2",
-    "which-module": "^1.0.0",
+    "string-width": "^2.0.0",
+    "which-module": "^2.0.0",
     "y18n": "^3.2.1",
-    "yargs-parser": "^4.2.0"
+    "yargs-parser": "^9.0.2"
   },
-  "deprecated": false,
   "description": "yargs the modern, pirate-themed, successor to optimist.",
   "devDependencies": {
-    "chai": "^3.4.1",
+    "chai": "^4.1.2",
     "chalk": "^1.1.3",
     "coveralls": "^2.11.11",
     "cpr": "^2.0.0",
@@ -52,14 +55,15 @@
     "es6-promise": "^4.0.2",
     "hashish": "0.0.4",
     "mocha": "^3.0.1",
-    "nyc": "^10.0.0",
+    "nyc": "^11.2.1",
     "rimraf": "^2.5.0",
     "standard": "^8.6.0",
-    "standard-version": "^3.0.0",
-    "which": "^1.2.9"
+    "standard-version": "^4.2.0",
+    "which": "^1.2.9",
+    "yargs-test-extends": "^1.0.1"
   },
   "engine": {
-    "node": ">=0.10"
+    "node": ">=4"
   },
   "files": [
     "index.js",
@@ -69,13 +73,6 @@
     "completion.sh.hbs",
     "LICENSE"
   ],
-  "greenkeeper": {
-    "ignore": [
-      "string-width",
-      "read-pkg-up",
-      "camelcase"
-    ]
-  },
   "homepage": "http://yargs.js.org/",
   "keywords": [
     "argument",
@@ -95,7 +92,7 @@
   },
   "scripts": {
     "coverage": "nyc report --reporter=text-lcov | coveralls",
-    "pretest": "standard",
+    "posttest": "standard",
     "release": "standard-version",
     "test": "nyc --cache mocha --require ./test/before.js --timeout=8000 --check-leaks"
   },
@@ -104,5 +101,5 @@
       "**/example/**"
     ]
   },
-  "version": "6.6.0"
+  "version": "11.0.0"
 }
