@@ -84,46 +77,45 @@
     "spdy": "^3.4.1",
     "strip-ansi": "^3.0.0",
     "supports-color": "^5.1.0",
-    "webpack-dev-middleware": "1.12.2",
-    "yargs": "6.6.0"
+    "webpack-dev-middleware": "3.1.3",
+    "webpack-log": "^1.1.2",
+    "yargs": "11.0.0"
   },
-  "deprecated": false,
   "description": "Serves a webpack app. Updates the browser on changes.",
   "devDependencies": {
     "babel-cli": "^6.26.0",
     "babel-core": "^6.26.0",
-    "babel-loader": "^7.1.2",
+    "babel-loader": "^7.1.4",
     "babel-preset-env": "^1.6.1",
     "codecov.io": "^0.1.6",
-    "copy-webpack-plugin": "^4.3.1",
-    "css-loader": "^0.28.5",
-    "eslint": "^4.5.0",
+    "copy-webpack-plugin": "^4.5.1",
+    "css-loader": "^0.28.10",
+    "eslint": "^4.18.2",
     "eslint-config-webpack": "^1.2.5",
-    "eslint-plugin-import": "^2.7.0",
-    "execa": "^0.8.0",
-    "file-loader": "^1.1.6",
-    "html-webpack-plugin": "^2.30.1",
+    "eslint-plugin-import": "^2.9.0",
+    "execa": "^0.9.0",
+    "file-loader": "^1.1.11",
+    "html-loader": "^0.5.5",
+    "html-webpack-plugin": "^3.0.6",
     "istanbul": "^0.4.5",
     "jquery": "^3.2.1",
-    "less": "^2.5.1",
-    "less-loader": "^4.0.5",
-    "marked": "^0.3.9",
+    "less": "^3.0.1",
+    "less-loader": "^4.1.0",
+    "marked": "^0.3.17",
     "mocha": "^3.5.3",
     "mocha-sinon": "^2.0.0",
-    "pug": "^2.0.0-beta5",
-    "pug-loader": "^2.3.0",
     "semver": "^5.4.1",
     "should": "^13.2.0",
-    "sinon": "^4.1.3",
-    "style-loader": "^0.19.1",
+    "sinon": "^4.4.2",
+    "style-loader": "^0.20.3",
     "supertest": "^3.0.0",
-    "uglifyjs-webpack-plugin": "^1.0.0-beta.2",
     "url-loader": "^0.6.2",
-    "webpack": "^3.10.0",
-    "ws": "^4.0.0"
+    "webpack": "^4.1.1",
+    "webpack-cli": "^2.0.14",
+    "ws": "^4.1.0"
   },
   "engines": {
-    "node": ">=4.7"
+    "node": ">=6.11.5"
   },
   "files": [
     "lib/",
@@ -143,7 +135,7 @@
   ],
   "name": "webpack-dev-server",
   "peerDependencies": {
-    "webpack": "^2.2.0 || ^3.0.0"
+    "webpack": "^4.0.0-beta.1"
   },
   "repository": {
     "type": "git",
