@@ -1,5 +1,5 @@
-module.exports = {
+export default {
   clear: 'both',
-  
+
   // border: '.5px solid black'
 }
